When it comes to optimizing code, probably the simplest and most effective way to optimize it is to count the number of instructions (yes, instructions; NOT lines of code) the code execute to achieve a certain goal. If the number of instructions can be reduced, there's a strong probability the code will perform faster.

Keeping that in mind,  this PR reduces the number of instructions on PushBack from 7 to 5 on most [branched executions](https://en.wikipedia.org/wiki/Branch_(computer_science)) (when d.tp is somewhere in the middle of the slice). The overall performance gain is significant. See below.

```
benchstat testdata/BenchmarkMicroserviceDequeQueueMaster.txt testdata/BenchmarkMicroserviceDequeQueuePR.txt
name        old time/op    new time/op    delta
/1-4           541ns ± 5%     506ns ± 4%  -6.51%  (p=0.000 n=10+10)
/10-4         3.71µs ± 4%    3.68µs ±12%    ~     (p=0.315 n=10+10)
/100-4        24.7µs ± 3%    24.6µs ± 4%    ~     (p=0.563 n=10+9)
/1000-4        245µs ± 4%     233µs ± 9%  -5.02%  (p=0.010 n=10+9)
/10000-4      2.50ms ± 3%    2.32ms ± 0%  -7.01%  (p=0.000 n=9+9)
/100000-4     27.4ms ± 4%    25.0ms ± 1%  -8.84%  (p=0.000 n=10+10)
/1000000-4     288ms ± 9%     273ms ± 6%  -5.48%  (p=0.019 n=10+10)

name        old alloc/op   new alloc/op   delta
/1-4            544B ± 0%      544B ± 0%    ~     (all equal)
/10-4         5.70kB ± 0%    5.70kB ± 0%    ~     (all equal)
/100-4        15.8kB ± 0%    15.8kB ± 0%    ~     (all equal)
/1000-4        137kB ± 0%     137kB ± 0%    ~     (all equal)
/10000-4      1.44MB ± 0%    1.44MB ± 0%    ~     (all equal)
/100000-4     14.4MB ± 0%    14.4MB ± 0%    ~     (p=0.211 n=10+10)
/1000000-4     144MB ± 0%     144MB ± 0%    ~     (all equal)

name        old allocs/op  new allocs/op  delta
/1-4            12.0 ± 0%      12.0 ± 0%    ~     (all equal)
/10-4           77.0 ± 0%      77.0 ± 0%    ~     (all equal)
/100-4           707 ± 0%       707 ± 0%    ~     (all equal)
/1000-4        7.02k ± 0%     7.02k ± 0%    ~     (all equal)
/10000-4       70.2k ± 0%     70.2k ± 0%    ~     (all equal)
/100000-4       702k ± 0%      702k ± 0%    ~     (all equal)
/1000000-4     7.02M ± 0%     7.02M ± 0%    ~     (all equal)

benchstat testdata/BenchmarkMicroserviceDequeStackMaster.txt testdata/BenchmarkMicroserviceDequeStackPR.txt
name        old time/op    new time/op    delta
/1-4           430ns ± 5%     402ns ± 6%  -6.42%  (p=0.000 n=10+10)
/10-4         2.76µs ± 6%    2.63µs ± 6%  -4.81%  (p=0.007 n=10+10)
/100-4        25.2µs ± 4%    24.5µs ± 3%  -2.91%  (p=0.022 n=10+9)
/1000-4        239µs ± 4%     236µs ± 5%    ~     (p=0.393 n=10+10)
/10000-4      2.42ms ± 4%    2.62ms ±18%    ~     (p=0.190 n=10+10)
/100000-4     27.0ms ±10%    25.0ms ± 1%  -7.12%  (p=0.000 n=10+10)
/1000000-4     281ms ± 5%     265ms ± 4%  -5.95%  (p=0.000 n=10+9)

name        old alloc/op   new alloc/op   delta
/1-4            288B ± 0%      288B ± 0%    ~     (all equal)
/10-4         1.55kB ± 0%    1.55kB ± 0%    ~     (all equal)
/100-4        15.8kB ± 0%    15.8kB ± 0%    ~     (all equal)
/1000-4        129kB ± 0%     129kB ± 0%    ~     (all equal)
/10000-4      1.44MB ± 0%    1.44MB ± 0%    ~     (all equal)
/100000-4     14.4MB ± 0%    14.4MB ± 0%    ~     (p=0.173 n=10+9)
/1000000-4     144MB ± 0%     144MB ± 0%    ~     (p=0.211 n=10+10)

name        old allocs/op  new allocs/op  delta
/1-4            11.0 ± 0%      11.0 ± 0%    ~     (all equal)
/10-4           75.0 ± 0%      75.0 ± 0%    ~     (all equal)
/100-4           707 ± 0%       707 ± 0%    ~     (all equal)
/1000-4        7.01k ± 0%     7.01k ± 0%    ~     (all equal)
/10000-4       70.2k ± 0%     70.2k ± 0%    ~     (all equal)
/100000-4       702k ± 0%      702k ± 0%    ~     (all equal)
/1000000-4     7.02M ± 0%     7.02M ± 0%    ~     (all equal)
```

I know, the code now looks worse. There's some replication as well. I keep making an unamused face every time I look at the PR code, but performance and efficiency comes before code cleanness. I'm sure most deque users doesn't really care how deque internal code looks like. Most of them care about deque being correct, performant and efficient. If we can do a better job on these areas by making the code looking a bit worse, we should do it. . :-)