@christianrpetrin 2018-12-01
Also increased maxSpareLinks from 4 to 5. The tests show good performance gains across all tests.

```
benchstat testdata/BenchmarkMicroserviceDequeQueueMaster.txt testdata/BenchmarkMicroserviceDequeQueuePR..txt
name        old time/op    new time/op    delta
/1-4           541ns ± 5%     502ns ± 1%  -7.21%  (p=0.000 n=10+10)
/10-4         3.71µs ± 4%    3.50µs ± 1%  -5.51%  (p=0.000 n=10+10)
/100-4        24.7µs ± 3%    23.8µs ± 1%  -3.61%  (p=0.000 n=10+10)
/1000-4        245µs ± 4%     229µs ± 1%  -6.52%  (p=0.000 n=10+10)
/10000-4      2.50ms ± 3%    2.35ms ± 0%  -5.96%  (p=0.000 n=9+7)
/100000-4     27.4ms ± 4%    25.5ms ± 2%  -6.76%  (p=0.000 n=10+10)
/1000000-4     288ms ± 9%     269ms ± 2%  -6.80%  (p=0.000 n=10+10)

name        old alloc/op   new alloc/op   delta
/1-4            544B ± 0%      544B ± 0%    ~     (all equal)
/10-4         5.70kB ± 0%    5.70kB ± 0%    ~     (all equal)
/100-4        15.8kB ± 0%    15.8kB ± 0%    ~     (all equal)
/1000-4        137kB ± 0%     133kB ± 0%  -3.02%  (p=0.000 n=10+10)
/10000-4      1.44MB ± 0%    1.43MB ± 0%  -0.58%  (p=0.000 n=10+10)
/100000-4     14.4MB ± 0%    14.4MB ± 0%  -0.06%  (p=0.000 n=10+10)
/1000000-4     144MB ± 0%     144MB ± 0%  -0.01%  (p=0.002 n=8+10)

name        old allocs/op  new allocs/op  delta
/1-4            12.0 ± 0%      12.0 ± 0%    ~     (all equal)
/10-4           77.0 ± 0%      77.0 ± 0%    ~     (all equal)
/100-4           707 ± 0%       707 ± 0%    ~     (all equal)
/1000-4        7.02k ± 0%     7.01k ± 0%  -0.03%  (p=0.000 n=10+10)
/10000-4       70.2k ± 0%     70.2k ± 0%  -0.01%  (p=0.000 n=10+10)
/100000-4       702k ± 0%      702k ± 0%  -0.00%  (p=0.000 n=10+10)
/1000000-4     7.02M ± 0%     7.02M ± 0%  -0.00%  (p=0.000 n=10+10)

benchstat testdata/BenchmarkMicroserviceDequeStackMaster.txt testdata/BenchmarkMicroserviceDequeStackPR.txt
name        old time/op    new time/op    delta
/1-4           430ns ± 5%     392ns ± 1%  -8.91%  (p=0.000 n=10+10)
/10-4         2.76µs ± 6%    2.57µs ± 1%  -6.80%  (p=0.000 n=10+9)
/100-4        25.2µs ± 4%    24.5µs ± 7%    ~     (p=0.089 n=10+10)
/1000-4        239µs ± 4%     234µs ± 8%    ~     (p=0.218 n=10+10)
/10000-4      2.42ms ± 4%    2.47ms ± 7%    ~     (p=0.123 n=10+10)
/100000-4     27.0ms ±10%    26.5ms ± 5%    ~     (p=0.481 n=10+10)
/1000000-4     281ms ± 5%     286ms ± 8%    ~     (p=0.436 n=10+10)

name        old alloc/op   new alloc/op   delta
/1-4            288B ± 0%      288B ± 0%    ~     (all equal)
/10-4         1.55kB ± 0%    1.55kB ± 0%    ~     (all equal)
/100-4        15.8kB ± 0%    15.8kB ± 0%    ~     (all equal)
/1000-4        129kB ± 0%     129kB ± 0%    ~     (all equal)
/10000-4      1.44MB ± 0%    1.43MB ± 0%  -0.58%  (p=0.000 n=10+10)
/100000-4     14.4MB ± 0%    14.4MB ± 0%  -0.06%  (p=0.000 n=10+9)
/1000000-4     144MB ± 0%     144MB ± 0%  -0.01%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
/1-4            11.0 ± 0%      11.0 ± 0%    ~     (all equal)
/10-4           75.0 ± 0%      75.0 ± 0%    ~     (all equal)
/100-4           707 ± 0%       707 ± 0%    ~     (all equal)
/1000-4        7.01k ± 0%     7.01k ± 0%    ~     (all equal)
/10000-4       70.2k ± 0%     70.2k ± 0%  -0.01%  (p=0.000 n=10+10)
/100000-4       702k ± 0%      702k ± 0%  -0.00%  (p=0.000 n=10+10)
/1000000-4     7.02M ± 0%     7.02M ± 0%  -0.00%  (p=0.000 n=10+10)
```